requirements_3.txt
exifread==2.1.2
natsort==6.0.0
pillow==5.4.1
PyHamcrest==1.9.0
tkfilebrowser==2.3.1
